Update (2023.03.08)
29783: LA port of 8151413: os::allocation_granularity/page_size and friends return signed values
29782: Replace NULL with nullptr in os_cpu/linux_loongarch

+++ src/hotspot/cpu/loongarch/interp_masm_loongarch_64.cpp
@@ -830,7 +830,7 @@ void InterpreterMacroAssembler::lock_object(Register lock_reg) {
     // least significant 2 bits clear.
     // NOTE: the oopMark is in tmp_reg as the result of cmpxchg
     sub_d(tmp_reg, tmp_reg, SP);
-    li(AT, 7 - os::vm_page_size());
+    li(AT, 7 - (int)os::vm_page_size());
     andr(tmp_reg, tmp_reg, AT);
     // Save the test result, for recursive case, the result is zero
     st_d(tmp_reg, lock_reg, mark_offset);
